requirements.txt
scikit-learn>=1.8,<1.9
torch>=2.10,<2.11
torchvision>=0.25,<0.26
scipy>=1.16,<1.17
Pillow>=11.1,<12
